Change `parseRoute` to `GitPath`
It's slightly less awkward than object destructuring followed by .toString

Also this makes .toString automatic (which is kinda silly and probably won't last forever)
